Refactor: move CanvasPanelProps to shared canvasTypes
- Extract CanvasPanelProps interface from Tutorial.tsx into types/canvasTypes.ts
- Update Tutorial.tsx and TopSection.tsx to import from @/types/canvasTypes
